requirements.txt
speech_recognition==3.10.0
pyaudio==0.2.11
pydub==0.25.1
openai-whisper==20231117
torch==2.1.0
transformers==4.35.0
nltk==3.8.1
textblob==0.17.1
numpy==1.24.3
scipy==1.11.3
scikit-learn==1.3.0
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
librosa==0.10.1
soundfile==0.12.1
webrtcvad==2.0.10
gtts==2.4.0
pyttsx3==2.90
language-tool-python==2.7.1 